refactor: AWS VM 어댑터 역할 분리
